Fix voice mask TTS
